requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
pydantic
requests
huggingface_hub
